SPLICE-2029 fix IT issue for mem platform
